feat: stories 폴더에 theme폴더 및 파일 만들고 theme관련 스타일 분리
