+++ stories/link/NavLink.tsx
@@ -1,0 +1,14 @@
+import Link from 'next/link'
+
+export interface NavLinkProps{
+    url:string;
+    font?:string;
+}
+
+export const NavLink = ({url="/",font="Montserrat"}) => {
+    return (
+        <Link href={`${url}`}>
+            
+        </Link>
+    )
+}
