+++ stories/title/Title.tsx
@@ -1,11 +1,6 @@
 import React from 'react'
-import { css } from '@emotion/react';
+import {fontTheme} from '@story/theme/theme'
 
-const fontTheme={
-    montserrat:css`font-family: 'Montserrat', sans-serif;`,
-    almendra:css`font-family: 'Almendra', serif;`,
-    nanumgothic:css`font-family: 'Nanum Gothic', sans-serif;`
-}
 
 export interface TitleProps{
     font?:'montserrat'|'almendra'|'nanumgothic';
